Finish renaming covidr -> aspen
### Description
covidr is a dumb name.

### Test plan
pip install -e src/py
