+++ .github/workflows/py-tests.yml
@@ -70,7 +70,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r src/py/requirements-dev.txt
-      - name: Install covidr
+      - name: Install aspen
         run: |
           pip install -e src/py
       - name: Run with pytest
