+++ Makefile
@@ -5,10 +5,10 @@ PYTHON_CODE_DIRECTORIES = src/py
 ### DATABASE #################################################
 #
 
-LOCAL_DB_CONTAINER_NAME = covidr-local
+LOCAL_DB_CONTAINER_NAME = aspen-local
 LOCAL_DB_CONTAINER_ID = $(shell docker ps -a | grep $(LOCAL_DB_CONTAINER_NAME) | awk '{print $$1}')
 LOCAL_DB_CONTAINER_RUNNING_ID = $(shell docker ps | grep $(LOCAL_DB_CONTAINER_NAME) | awk '{print $$1}')
-LOCAL_DB_NAME = covidr_db
+LOCAL_DB_NAME = aspen_db
 LOCAL_DB_ADMIN_USERNAME = postgres  # This has to be "postgres" to ease moving snapshots from RDS.
 LOCAL_DB_ADMIN_PASSWORD = admin
 LOCAL_DB_RW_USERNAME = user_rw
@@ -37,7 +37,7 @@ setup-local-db:
 
 init-local-db:
 	@$(MAKE) setup-local-db
-	covidr-cli db create
+	aspen-cli db create
 	ENV=dev alembic stamp head
 
 stop-local-db:
