+++ README.md
@@ -1,4 +1,4 @@
-# COVIDR
+# ASPEN
 
 This repository contains the code for the gen-epi project.
 
@@ -6,11 +6,11 @@ This repository contains the code for the gen-epi project.
 ## Running the app in development mode: 
 
 ```bash
-covidr% npm --prefix src/ts start
-covidr% export FLASK_APP=covidr.app
-covidr% export FLASK_ENV=development
-covidr% cd src/py
-covidr% flask run --host localhost --port 3000  # host and port needed for auth0
+aspen% npm --prefix src/ts start
+aspen% export FLASK_APP=aspen.app
+aspen% export FLASK_ENV=development
+aspen% cd src/py
+aspen% flask run --host localhost --port 3000  # host and port needed for auth0
 ```
 
 ## Backend
