+++ database_migrations/versions/20212415_112447_add_group_and_user_models.py
@@ -23,7 +23,7 @@ def upgrade():
         sa.Column("name", sa.String(), nullable=False),
         sa.PrimaryKeyConstraint("id", name=op.f("pk_groups")),
         sa.UniqueConstraint("name", name=op.f("uq_groups_name")),
-        schema="covidr",
+        schema="aspen",
     )
     op.create_table(
         "users",
@@ -35,20 +35,20 @@ def upgrade():
         sa.Column("group_id", sa.Integer(), nullable=False),
         sa.ForeignKeyConstraint(
             ["group_id"],
-            ["covidr.groups.id"],
+            ["aspen.groups.id"],
             name=op.f("fk_users_group_id_groups"),
         ),
         sa.PrimaryKeyConstraint("id", name=op.f("pk_users")),
         sa.UniqueConstraint(
             "auth0_user_id", name=op.f("uq_users_auth0_user_id")
         ),
-        schema="covidr",
+        schema="aspen",
     )
     # ### end Alembic commands ###
 
 
 def downgrade():
     # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("users", schema="covidr")
-    op.drop_table("groups", schema="covidr")
+    op.drop_table("users", schema="aspen")
+    op.drop_table("groups", schema="aspen")
     # ### end Alembic commands ###
