+++ src/py/aspen/cli/db.py
@@ -1,7 +1,7 @@
 import click
-from covidr.config.development import DevelopmentConfig
-from covidr.database.connection import get_db_uri, init_db
-from covidr.database.schema import create_tables_and_schema
+from aspen.config.development import DevelopmentConfig
+from aspen.database.connection import get_db_uri, init_db
+from aspen.database.schema import create_tables_and_schema
 
 from .toplevel import cli
