+++ src/py/aspen/config/development.py
@@ -14,7 +14,7 @@ class DevelopmentConfig(Config, descriptive_name="dev"):
 class DevelopmentDatabaseConfig(DatabaseConfig):
     @property
     def URI(self):
-        return "postgresql://user_rw:password_rw@localhost:5432/covidr_db"
+        return "postgresql://user_rw:password_rw@localhost:5432/aspen_db"
 
     @property
     def SEND_FILE_MAX_AGE_DEFAULT(self):
