+++ src/ts/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "covidr-ts-frontend",
+  "name": "aspen-ts-frontend",
   "version": "0.1.0",
   "private": true,
   "proxy": "http://localhost:5000",
