+++ src/ts/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "covidr",
-  "name": "covidr",
+  "short_name": "aspen",
+  "name": "aspen",
   "icons": [
     {
       "src": "favicon.ico",
